rename root namespace to ru.nsu.aeliseev2.task111
