+++ Task_1_1_1/build.gradle
@@ -18,7 +18,7 @@ dependencies {
 }
 
 application {
-    mainClass = "ru.nsu.aeliseev2.task_1_1_1.App"
+    mainClass = "ru.nsu.aeliseev2.task111.App"
 }
 
 test {
